refactor(Order): 응용서비스계층 interface를 Order유형별로 정의
