--- src/main/java/kitchenpos/order/common/application/CreateOrder.java
@@ -1,8 +1,0 @@
-package kitchenpos.order.common.application;
-
-import kitchenpos.order.common.domain.Order;
-
-public interface CreateOrder {
-
-    Order create(Order order);
-}
