--- src/main/java/kitchenpos/order/common/application/ServeOrder.java
@@ -1,9 +1,0 @@
-package kitchenpos.order.common.application;
-
-import java.util.UUID;
-import kitchenpos.order.common.domain.Order;
-
-public interface ServeOrder {
-
-    Order serve(UUID orderId);
-}
