--- src/main/java/kitchenpos/order/deliveryorder/application/CompleteDelivery.java
@@ -1,9 +1,0 @@
-package kitchenpos.order.deliveryorder.application;
-
-import java.util.UUID;
-import kitchenpos.order.common.domain.Order;
-
-public interface CompleteDelivery {
-
-    Order completeDelivery(UUID deliveryId);
-}
