--- src/main/java/kitchenpos/order/deliveryorder/application/StartDelivery.java
@@ -1,9 +1,0 @@
-package kitchenpos.order.deliveryorder.application;
-
-import java.util.UUID;
-import kitchenpos.order.common.domain.Order;
-
-public interface StartDelivery {
-
-    Order startDelivery(UUID orderId);
-}
