+++ src/main/java/kitchenpos/order/takeoutorder/application/TakeoutOrderService.java
@@ -1,113 +1,15 @@
 package kitchenpos.order.takeoutorder.application;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.UUID;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.order.common.application.AcceptOrder;
-import kitchenpos.order.common.application.CompleteOrder;
-import kitchenpos.order.common.application.CreateOrder;
-import kitchenpos.order.common.application.ServeOrder;
 import kitchenpos.order.common.domain.Order;
-import kitchenpos.order.common.domain.OrderLineItem;
-import kitchenpos.order.common.domain.OrderRepository;
-import kitchenpos.order.common.domain.OrderStatus;
-import kitchenpos.order.common.domain.OrderType;
-import kitchenpos.order.takeoutorder.domain.TakeoutOrder;
-import org.springframework.stereotype.Service;
 
-@Service
-public class TakeoutOrderService implements CreateOrder, AcceptOrder, ServeOrder, CompleteOrder {
+public interface TakeoutOrderService {
 
-    private final OrderRepository orderRepository;
-    private final MenuRepository menuRepository;
+    Order create(final Order order);
 
-    public TakeoutOrderService(OrderRepository orderRepository, MenuRepository menuRepository) {
-        this.orderRepository = orderRepository;
-        this.menuRepository = menuRepository;
-    }
+    Order accept(final UUID orderId);
 
-    @Override
-    public Order create(Order request) {
-        final OrderType type = request.getType();
-        if (Objects.isNull(type)) {
-            throw new IllegalArgumentException();
-        }
-        final List<OrderLineItem> orderLineItemRequests = request.getOrderLineItems();
-        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-        final List<Menu> menus = menuRepository.findAllByIdIn(
-            orderLineItemRequests.stream()
-                .map(OrderLineItem::getMenuId)
-                .toList()
-        );
-        if (menus.size() != orderLineItemRequests.size()) {
-            throw new IllegalArgumentException();
-        }
-        final List<OrderLineItem> orderLineItems = new ArrayList<>();
-        for (final OrderLineItem orderLineItemRequest : orderLineItemRequests) {
-            final long quantity = orderLineItemRequest.getQuantity();
-            final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
-                .orElseThrow(NoSuchElementException::new);
-            if (!menu.isDisplayed()) {
-                throw new IllegalStateException();
-            }
-            if (menu.getPrice().compareTo(orderLineItemRequest.getPrice()) != 0) {
-                throw new IllegalArgumentException();
-            }
-            final OrderLineItem orderLineItem = new OrderLineItem();
-            orderLineItem.setMenu(menu);
-            orderLineItem.setQuantity(quantity);
-            orderLineItems.add(orderLineItem);
-        }
-        TakeoutOrder order = new TakeoutOrder();
-        order.setId(UUID.randomUUID());
-        order.setStatus(OrderStatus.WAITING);
-        order.setOrderDateTime(LocalDateTime.now());
-        order.setOrderLineItems(orderLineItems);
-        return orderRepository.save(order);
-    }
+    Order serve(final UUID orderId);
 
-    @Override
-    public Order accept(UUID orderId) {
-        final Order order = orderRepository.findById(orderId)
-            .orElseThrow(NoSuchElementException::new);
-        if (order.getStatus() != OrderStatus.WAITING) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.ACCEPTED);
-        return order;
-    }
-
-    @Override
-    public Order serve(UUID orderId) {
-        final Order order = orderRepository.findById(orderId)
-            .orElseThrow(NoSuchElementException::new);
-        if (order.getStatus() != OrderStatus.ACCEPTED) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.SERVED);
-        return order;
-    }
-
-    @Override
-    public Order complete(UUID orderId) {
-        final Order order = orderRepository.findById(orderId)
-            .orElseThrow(NoSuchElementException::new);
-        final OrderType type = order.getType();
-        final OrderStatus status = order.getStatus();
-        if (type == OrderType.TAKEOUT) {
-            if (status != OrderStatus.SERVED) {
-                throw new IllegalStateException();
-            }
-        }
-        order.setStatus(OrderStatus.COMPLETED);
-        return order;
-    }
-
+    Order complete(final UUID orderId);
 }
